[refactor] 개인액션 상세 - 채팅 버튼, 채팅 모달창 컴포넌트 분리
